TAN-580 New Parent class for Aws endpoints workers #comment Removed/added some properties on message files as well
